feat(coin-evm): abstract ChainProps
