+++ packages/coin-evm/tools/chain.ts
@@ -1,0 +1,4 @@
+import { CHAIN } from '../src';
+
+console.log('Cronos Chain Information', CHAIN.CRONOS.toHexChainInfo());
+console.log('Cronos Chain Signature', CHAIN.CRONOS.getSignature());
